fix: derive canonical links from base url
