+++ src/lib/canonical-developer-links.test.ts
@@ -1,11 +1,14 @@
 import { describe, expect, test } from 'vitest'
 import { canonicalizeGeneratedDeveloperLinks } from './canonical-developer-links'
 
+const publicDevelopersUrl = 'https://tempo.xyz/developers/docs'
+
 describe('canonicalizeGeneratedDeveloperLinks', () => {
   test('uses canonical public URLs for generated HTML hrefs', () => {
     expect(
       canonicalizeGeneratedDeveloperLinks(
         '<a href="/docs">Docs</a><a href="/docs/api#authentication">API</a>',
+        publicDevelopersUrl,
       ),
     ).toBe(
       '<a href="https://tempo.xyz/developers/docs">Docs</a><a href="https://tempo.xyz/developers/docs/api#authentication">API</a>',
@@ -16,6 +19,7 @@ describe('canonicalizeGeneratedDeveloperLinks', () => {
     expect(
       canonicalizeGeneratedDeveloperLinks(
         '{"href":"/docs/api","to":"/docs/api"} {\\"href\\":\\"/docs/api\\",\\"to\\":\\"/docs/api\\"}',
+        publicDevelopersUrl,
       ),
     ).toBe(
       '{"href":"https://tempo.xyz/developers/docs/api","to":"/docs/api"} {\\"href\\":\\"https://tempo.xyz/developers/docs/api\\",\\"to\\":\\"/docs/api\\"}',
@@ -31,6 +35,7 @@ describe('canonicalizeGeneratedDeveloperLinks', () => {
           '- [Authentication](/docs/api#authentication)',
           '<Card title="API" to="/docs/api" />',
         ].join('\n'),
+        publicDevelopersUrl,
       ),
     ).toBe(
       [
@@ -46,10 +51,20 @@ describe('canonicalizeGeneratedDeveloperLinks', () => {
     expect(
       canonicalizeGeneratedDeveloperLinks(
         '<a href="/developers/docs/api">API</a> [API](/developers/docs/api)',
+        publicDevelopersUrl,
       ),
     ).toBe(
       '<a href="https://tempo.xyz/developers/docs/api">API</a> [API](https://tempo.xyz/developers/docs/api)',
     )
+  })
+
+  test('uses the configured public URL', () => {
+    expect(
+      canonicalizeGeneratedDeveloperLinks(
+        '<a href="/docs/api">API</a>',
+        'https://docs.example.com/reference/docs',
+      ),
+    ).toBe('<a href="https://docs.example.com/reference/docs/api">API</a>')
   })
 
   test('leaves internal route values and unrelated URLs unchanged', () => {
@@ -59,6 +74,6 @@ describe('canonicalizeGeneratedDeveloperLinks', () => {
       '[External](https://example.com/docs)',
     ].join('\n')
 
-    expect(canonicalizeGeneratedDeveloperLinks(content)).toBe(content)
+    expect(canonicalizeGeneratedDeveloperLinks(content, publicDevelopersUrl)).toBe(content)
   })
 })
