+++ src/marketing/seo.ts
@@ -5,6 +5,8 @@
 import { OG_IMAGE_VERSION } from '../lib/og-sections'
 import { type CategorySlug, categoryBySlug } from './app/blog/_lib/categories'
 
+export { resolveBaseUrl } from '../lib/base-url'
+
 export type PostSeo = {
   slug: string
   title: string // raw post title (no " — Tempo Developers" suffix)
@@ -12,20 +14,6 @@ export type PostSeo = {
   date: string // YYYY-MM-DD
   category: CategorySlug
   authors: string
-}
-
-// Mirrors the baseUrl resolution in vocs.config.ts so canonical and OG URLs are
-// absolute in production and gracefully relative on preview/local builds
-// (returning '' there, just like the docs site, to avoid leaking preview URLs).
-export function resolveBaseUrl(): string {
-  if (process.env.VERCEL_ENV && process.env.VERCEL_ENV !== 'production') return ''
-  if (URL.canParse(process.env.VITE_BASE_URL ?? '')) {
-    return (process.env.VITE_BASE_URL as string).replace(/\/$/, '')
-  }
-  if (process.env.VERCEL_ENV === 'production') return 'https://tempo.xyz/developers'
-  const productionUrl = process.env.VERCEL_PROJECT_PRODUCTION_URL
-  if (productionUrl) return `https://${productionUrl}`
-  return ''
 }
 
 export function absoluteUrl(base: string, pathname: string): string {
